refactor: colocate satisfiability unit tests with their modules
Following the file split, the giant `tests.rs` is broken up so each
module owns its own `#[cfg(test)] mod tests { ... }`:

- `pypi.rs`: `pypi_satisfies_*` checks, index-mismatch regressions,
  `PypiNoBuildCheck` regression test, plus the `lock_for_test` helper.
- `source_record.rs`: the entire `backend_verification` test set
  (`variants_equivalent`, `verify_locked_against_backend_specs`,
  `diff_dep_sequences`, `verify_locked_run_deps_against_backend`,
  `build_full_source_record_from_output`).
- `tests.rs`: kept for the cross-cutting orchestration tests
  (`test_good_satisfiability`, `test_failing_satisfiability`, `q`)
  plus the `verify_lockfile_satisfiability` harness and
  `LockfileUnsat` shim. Module path is preserved so the existing
  `satisfiability/snapshots/*.snap` files keep matching.

With tests living next to their code, items previously promoted to
`pub(crate)` so an out-of-module `tests` could see them
(`variants_equivalent`, `diff_dep_sequences`,
`verify_locked_against_backend_specs`,
`verify_locked_run_deps_against_backend`,
`build_full_source_record_from_output`, `DepDiff`) revert to private
visibility — children of a module can see its private items.
The `#[cfg(test)] pub(super) use source_record::*` trampoline in
`mod.rs` is no longer needed.

127 satisfiability tests pass against the new layout.
